feat: StreamUtils renamed to AudioStreamUtils and now is public
